Differential Expression with Dream

Differential expression testing is performed with "variancePartition::dream()" to increase power and decrease false positives for repeated measure designs. The `primary_variable` is modeled as a fixed effect. If you wish to test an interaction term, `primary_variable` can take multiple variable names.

differential_expression(
  filtered_counts,
  cqn_counts,
  md,
  primary_variable,
  biomart_results,
  p_value_threshold,
  fold_change_threshold,
  model_variables = NULL,
  exclude_variables = NULL,
  cores = NULL
)

Arguments

filtered_counts

A counts data frame with genes removed that have low expression.

cqn_counts

A counts data frame normalized by CQN.

md

A data frame with sample identifiers in a column and relevant experimental covariates.

primary_variable

Vector of variables that will be collapsed into a single fixed effect interaction term.

biomart_results

Output of "sageseqr::get_biomart()". Gene Ids are stored as rownames.

p_value_threshold

Numeric. P-values are adjusted by Benjamini and Hochberg (BH) false discovery rate (FDR). Significant genes are those with an adjusted p-value greater than this threshold.

fold_change_threshold

Numeric. Significant genes are those with a fold-change greater than this threshold.

model_variables

Optional. Vector of variables to include in the linear (mixed) model. If not supplied, the model will include all variables in md.

exclude_variables

Vector of variables to exclude from testing.

cores

An integer of cores to specify in the parallel backend (eg. 4).

Value

A named list with "variancePartition::voomWithDreamWeights()" normalized counts, contrasts from "variancePartition::getContrasts()", linear mixed model fits from "variancePartition::dream()", differential expression results from "limma::topTable()" and gene feature-specific metadata, the response variable, the model formula fit to compute differential expression results. The list names are: "list(voom_object, contrasts_to_plot, fits, differential_expression, primary_variable, formula)"