Execute the drake RNA-seq plan

This function wraps the "drake::plan()" and copies the R markdown report to the user's working directory.

rnaseq_plan(
  metadata_id,
  metadata_version,
  counts_id,
  counts_version,
  gene_id_input,
  sample_id_input,
  factor_input,
  continuous_input,
  biomart_id,
  biomart_version,
  host,
  filters,
  organism,
  conditions,
  cpm_threshold = 1,
  conditions_threshold = 0.5,
  primary_variable,
  sex_var,
  color,
  shape,
  size,
  report_name,
  skip_model,
  de_contrasts,
  fold_change_threshold,
  p_value_threshold,
  parent_id,
  rownames,
  config_file,
  gene_list,
  force_model = NULL
)

Arguments

metadata_id
Synapse ID to clean metadata file with sample identifiers in a column and variables of interest as column names. There cannot be any missing values.

metadata_version
Optionally, include Synapse file version number. If omitted, current version will be downloaded.

counts_id
Synapse ID to counts data frame with identifiers to the metadata as column names and gene ids in a column.

counts_version
Optionally, include Synapse file version number.

gene_id_input
Column name of the gene ids in the counts_id file.

sample_id_input
Column name of the sample ids in the metadata_id file.

factor_input
Vector of factor variables. Variables must be present in the metadata as column names.

continuous_input
Vector of continuous variables. Variables must be present in the metadata as column names.

biomart_id
Synapse ID to biomart object.

biomart_version
Optionally, include Synapse file version number.

host
An optional character vector specifying the release version. This specification is highly recommended for a reproducible workflow. (see "biomaRt::listEnsemblArchives()")

filters
A character vector listing biomaRt query filters. (For a list of filters see "biomaRt::listFilters()")

organism
A character vector of the organism name. This argument takes partial strings. For example,"hsa" will match "hsapiens_gene_ensembl".

conditions
Optional. Conditions to bin gene counts that correspond to variables in `md`.

cpm_threshold
The minimum number of CPM allowed.

conditions_threshold
Percentage of samples that should contain the minimum CPM.

primary_variable
Baseline variable for model selection and variable to stratify groups in the boxplot.

sex_var
Column name of the sex or gender-specific metadata.

color
Discrete variable in `clean_metadata` differentiated by color.

shape
Discrete variable in `clean_metadata` differentiated by shape.

size
Continuous variable in `clean_metadata` differentiated by size.

report_name
Name of output markdown file.

skip_model
If TRUE, does not run regression model.

de_contrasts
Required. Variables in the metadata to define comparisons between groups.

fold_change_threshold
Numeric. Significant genes are those with a fold-change greater than this threshold.

p_value_threshold
Numeric. P-values are adjusted by Benjamini and Hochberg (BH) false discovery rate (FDR). Significant genes are those with an adjusted p-value greater than this threshold.

parent_id
A Synapse Id that corresponds to a project or folder to store output.

rownames
A list of variables to store rownames ordered by `metadata`, `filtered_counts`, `biomart_results`, `cqn_counts`. If not applicable, set as NULL.

config_file
Optional. Path to configuration file.

gene_list
Optional. A vector of genes to label in the volcano plot.

force_model
Optional. A vector of variables to include in the differential expression model if you want to skip the output of the stepwise regression.