Store output to Synapse

Store transformed data, markdown and html report to a Folder in Synapse.

store_results(
  clean_md = clean_md,
  filtered_counts = filtered_counts,
  biomart_results = biomart_results,
  cqn_counts = cqn_counts$E,
  de_results = de,
  residualized_counts = residualized_counts,
  report = report,
  syn_names,
  data_names,
  parent_id,
  inputs,
  activity_provenance,
  rownames = NULL,
  config_file = NULL,
  report_name = NULL
)

Arguments

clean_md

The target containing the metadata data frame. Defaults to target name constrained by "targets::tar_make".

filtered_counts

The target containing counts after low gene expression has been removed. Defaults to target name constrained by "targets::tar_make()".

biomart_results

The target containing gene annotations from biomart. Defaults to target name constrained by "targets::tar_make".

cqn_counts

The target containing Conditional Quantile Normalized (CQN) counts. Defaults to target name constrained by "targets::tar_make".

de_results

The target containing differential expression gene lists. Defaults to target name constrained in the _targets.R file.

residualized_counts

The target containing counts adjusted for batch effects. Defaults to target name constrained in the _targets.R file.

report

The target containing the rendered html document. Defaults to target name constrained in the _targets.R file.

syn_names

A list of human-readable names for the Synapse entities ordered by `metadata`, `filtered_counts`, `biomart_results`, `cqn_counts`.

data_names

A list of identifiers to embed in the file name ordered by `clean_md`, `filtered_counts`, `biomart_results`, `cqn_counts`, `de_results`.

parent_id

A Synapse Id that corresponds to a project or folder to store output.

inputs

A character vector of Synapse Ids to create provenance between output files and input files.

activity_provenance

A phrase to describe the data transformation for provenance.

rownames

A list of variables to store rownames ordered by `metadata`, `filtered_counts`, `biomart_results`, `cqn_counts`. If not applicable, set as NULL.

config_file

Optional. Path to configuration file.

report_name

Name of output markdown file.